+++ src/EFPListRegistry.sol
@@ -16,7 +16,7 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
     ///////////////////////////////////////////////////////////////////////////
 
     /// @notice Emitted when the mint batch size is changed.
-    event MaxMintBatchSizeChange(uint maxMintBatchSize);
+    event MaxMintBatchSizeChange(uint256 maxMintBatchSize);
 
     /// @notice Emitted when the mint state is changed.
     event MintStateChange(MintState mintState);
@@ -32,14 +32,14 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
     MintState private mintState = MintState.Disabled;
 
     /// @notice The maximum number of tokens that can be minted in a single batch.
-    uint private maxMintBatchSize = 10000;
+    uint256 private maxMintBatchSize = 10000;
 
     /// @notice The price oracle. If set, the price oracle is used to determine
     /// the price of minting.
     IEFPListPriceOracle private priceOracle;
 
     /// @notice The list storage location associated with a token.
-    mapping(uint => bytes) private tokenIdToListStorageLocation;
+    mapping(uint256 => bytes) private tokenIdToListStorageLocation;
 
     ///////////////////////////////////////////////////////////////////////////
     // Constructor
@@ -71,7 +71,7 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
     ///////////////////////////////////////////////////////////////////////////
 
     /// @notice Restrict access to the owner of a specific token.
-    modifier onlyTokenOwner(uint tokenId) {
+    modifier onlyTokenOwner(uint256 tokenId) {
         require(ownerOf(tokenId) == msg.sender, "EFP: caller is not the owner");
         _;
     }
@@ -89,10 +89,7 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
     modifier mintBatchAllowed() {
         require(mintState != MintState.Disabled, "EFP: minting is disabled");
         require(mintState != MintState.OwnerOnly || msg.sender == owner(), "EFP: minting is restricted to owner");
-        require(
-            mintState != MintState.PublicMint || msg.sender == owner(),
-            "EFP: batch minting is restricted to owner"
-        );
+        require(mintState != MintState.PublicMint || msg.sender == owner(), "EFP: batch minting is restricted to owner");
         // else PublicBatch allowed
         _;
     }
@@ -106,7 +103,7 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
      * @param tokenId The ID of the token.
      * @return The list location.
      */
-    function getListStorageLocation(uint tokenId) external view override returns (bytes memory) {
+    function getListStorageLocation(uint256 tokenId) external view override returns (bytes memory) {
         return tokenIdToListStorageLocation[tokenId];
     }
 
@@ -115,10 +112,11 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
      * @param tokenId The ID of the token.
      * @param listStorageLocation The list storage location to be associated with the token.
      */
-    function setListStorageLocation(
-        uint tokenId,
-        bytes calldata listStorageLocation
-    ) external override onlyTokenOwner(tokenId) {
+    function setListStorageLocation(uint256 tokenId, bytes calldata listStorageLocation)
+        external
+        override
+        onlyTokenOwner(tokenId)
+    {
         _setListStorageLocation(tokenId, listStorageLocation);
     }
 
@@ -127,7 +125,7 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
      * @param tokenId The ID of the token.
      * @param listStorageLocation The list storage location to be associated with the token.
      */
-    function _setListStorageLocation(uint tokenId, bytes calldata listStorageLocation) internal {
+    function _setListStorageLocation(uint256 tokenId, bytes calldata listStorageLocation) internal {
         tokenIdToListStorageLocation[tokenId] = listStorageLocation;
         emit ListStorageLocationChange(tokenId, listStorageLocation);
     }
@@ -149,13 +147,13 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
     }
 
     /// @notice Fetches the max mint batch size.
-    function getMaxMintBatchSize() external view returns (uint) {
+    function getMaxMintBatchSize() external view returns (uint256) {
         return maxMintBatchSize;
     }
 
     /// @notice Sets the max mint batch size.
     /// @param _maxMintBatchSize The new max mint batch size.
-    function setMaxMintBatchSize(uint _maxMintBatchSize) external onlyOwner {
+    function setMaxMintBatchSize(uint256 _maxMintBatchSize) external onlyOwner {
         maxMintBatchSize = _maxMintBatchSize;
         emit MaxMintBatchSizeChange(_maxMintBatchSize);
     }
@@ -165,8 +163,8 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
      * @param listStorageLocation The list storage location to be associated with the token.
      */
     function mint(bytes calldata listStorageLocation) public payable mintAllowed {
-        uint tokenId = totalSupply();
-        uint price = (address(priceOracle) != address(0)) ? priceOracle.getPrice(tokenId, 1) : 0;
+        uint256 tokenId = totalSupply();
+        uint256 price = (address(priceOracle) != address(0)) ? priceOracle.getPrice(tokenId, 1) : 0;
         require(msg.value >= price, "insufficient funds");
 
         _safeMint(msg.sender, 1);
@@ -179,8 +177,8 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
      * @param listStorageLocation The list storage location to be associated with the token.
      */
     function mintTo(address to, bytes calldata listStorageLocation) public payable mintAllowed {
-        uint tokenId = totalSupply();
-        uint price = (address(priceOracle) != address(0)) ? priceOracle.getPrice(tokenId, 1) : 0;
+        uint256 tokenId = totalSupply();
+        uint256 price = (address(priceOracle) != address(0)) ? priceOracle.getPrice(tokenId, 1) : 0;
         require(msg.value >= price, "insufficient funds");
 
         _safeMint(to, 1);
@@ -189,10 +187,10 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
 
     /// @notice Mints a batch of new tokens.
     /// @param quantity The number of tokens to mint.
-    function mintBatch(uint quantity) public payable mintBatchAllowed {
+    function mintBatch(uint256 quantity) public payable mintBatchAllowed {
         require(quantity <= maxMintBatchSize, "batch size too big");
 
-        uint price = (address(priceOracle) != address(0)) ? priceOracle.getPrice(totalSupply(), quantity) : 0;
+        uint256 price = (address(priceOracle) != address(0)) ? priceOracle.getPrice(totalSupply(), quantity) : 0;
         require(msg.value >= price, "insufficient funds");
 
         _safeMint(msg.sender, quantity);
@@ -202,10 +200,10 @@ contract EFPListRegistry is IEFPListRegistry, ERC721A, Ownable {
     /// @notice Mints a batch of new tokens.
     /// @param to The address to mint the tokens to.
     /// @param quantity The number of tokens to mint.
-    function mintBatchTo(address to, uint quantity) public payable mintBatchAllowed {
+    function mintBatchTo(address to, uint256 quantity) public payable mintBatchAllowed {
         require(quantity <= maxMintBatchSize, "batch size too big");
 
-        uint price = (address(priceOracle) != address(0)) ? priceOracle.getPrice(totalSupply(), quantity) : 0;
+        uint256 price = (address(priceOracle) != address(0)) ? priceOracle.getPrice(totalSupply(), quantity) : 0;
         require(msg.value >= price, "insufficient funds");
 
         _safeMint(to, quantity);
