+++ src/interfaces/IEFPListPriceOracle.sol
@@ -5,5 +5,5 @@ pragma solidity ^0.8.20;
  * @title IEFPListPriceOracle
  */
 interface IEFPListPriceOracle {
-    function getPrice(uint tokenId, uint quantity) external view returns (uint);
+    function getPrice(uint256 tokenId, uint256 quantity) external view returns (uint256);
 }
